pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "gunicorn>=25.3.0",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "werkzeug>=3.1.8",
]
